refactor: rename electron modules to be more explicit and to differentiate between Electron window and the global window variable in renderer worker
