+++ packages/main-process/src/parts/App/App.js
@@ -3,14 +3,14 @@ const { MessageChannel } = require('worker_threads')
 const { app } = require('electron')
 const unhandled = require('electron-unhandled') // TODO this might slow down initial startup
 const minimist = require('minimist')
-const Electron = require('../Electron/Electron.js')
+const Electron = require('electron')
 const Platform = require('../Platform/Platform.js')
 const SharedProcess = require('../SharedProcess/SharedProcess.js')
 const Debug = require('../Debug/Debug.js')
 const LifeCycle = require('../LifeCycle/LifeCycle.js')
 const Performance = require('../Performance/Performance.js')
 const Cli = require('../Cli/Cli.js')
-const Session = require('../Session/Session.js')
+const Session = require('../ElectronSession/ElectronSession.js')
 const AppWindow = require('../AppWindow/AppWindow.js')
 
 // TODO use Platform.getScheme() instead of Product.getTheme()
