--- packages/main-process/src/parts/Electron/Electron.js
@@ -1,25 +1,0 @@
-const {
-  BrowserWindow,
-  app,
-  ipcMain,
-  screen,
-  dialog,
-  protocol,
-  session,
-  shell,
-  clipboard,
-  Menu,
-} = require('electron')
-
-module.exports = {
-  BrowserWindow,
-  app,
-  screen,
-  dialog,
-  protocol,
-  session,
-  shell,
-  clipboard,
-  ipcMain,
-  Menu,
-}
