+++ packages/main-process/src/parts/ElectronBeep/ElectronBeep.ipc.js
@@ -1,4 +1,4 @@
-const Beep = require('./Beep.js')
+const Beep = require('./ElectronBeep.js')
 
 exports.Commands = {
   'Beep.beep': Beep.beep,
