+++ packages/main-process/src/parts/ElectronBeep/ElectronBeep.js
@@ -1,5 +1,7 @@
 const { shell } = require('electron')
 
+// TODO rename module to ElectronShell
+
 exports.beep = () => {
   shell.beep()
 }
