+++ packages/main-process/src/parts/ElectronDeveloper/ElectronDeveloper.js
@@ -1,6 +1,4 @@
 const { MessageChannelMain, ipcMain } = require('electron')
-const { join } = require('path')
-const { BrowserWindow } = require('../Electron/Electron.js')
 const Performance = require('../Performance/Performance.js')
 
 exports.getPerformanceEntries = () => {
