+++ packages/main-process/src/parts/ElectronSession/ElectronSession.js
