+++ packages/main-process/src/parts/ElectronWindowProcessExplorer/ElectronWindowProcessExplorer.js
@@ -1,10 +1,10 @@
 const { ipcMain } = require('electron')
 const { writeFile } = require('fs/promises')
 const VError = require('verror')
-const { BrowserWindow, Menu } = require('../Electron/Electron.js')
+const { BrowserWindow, Menu } = require('electron')
 const ListProcessesWithMemoryUsage = require('../ListProcessesWithMemoryUsage/ListProcessesWithMemoryUsage.js')
 const Platform = require('../Platform/Platform.js')
-const Session = require('../Session/Session.js')
+const Session = require('../ElectronSession/ElectronSession.js')
 const ColorTheme = require('../ColorTheme/ColorTheme.js')
 const Path = require('../Path/Path.js')
 const Root = require('../Root/Root.js')
