+++ packages/main-process/test/Window.test.js
@@ -1,4 +1,4 @@
-const Window = require('../src/parts/Window/Window.js')
+const Window = require('../src/parts/ElectronWindow/ElectronWindow.js')
 
 // TODO how to test window (highly depends on electron)
 test.skip('toggleDevtools', () => {
